tracking is working and style refactoring

+++ shared/src/models/prices.rs
@@ -4,14 +4,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Default)]
 pub struct CityPrice {
-    pub sell_price_min: u64,
-    pub buy_price_max: u64,
+    pub sell_price_min: u32,
+    pub buy_price_max: u32,
     pub updated_at: DateTime<Utc>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Default)]
 pub struct ItemPrice {
-    pub current: u64,
+    pub current: u32,
     pub cities: HashMap<String, CityPrice>,
 }
